Pass a resource directory to the graphics engine instead of hardcoding it and move everything from bin/x86/Data/ to content/
Physics is currently the only place I'm aware of that's still hardcoding the resource path

+++ src/skyflower/LineChart.cpp
@@ -102,7 +102,7 @@ void LineChart::Draw(float startTime, float endTime, float resolution, float tar
 	float maxValue = FLT_MIN;
 	float minValue = FLT_MAX;
 	double totalValue = 0.0;
-	int previousIndex;
+	size_t previousIndex;
 	int count = 0;
 	bool first = true;
